requirements.txt
Django==1.5.1
Pillow==2.0.0
South==0.8.1
django-compressor==1.3
https://github.com/deschler/django-modeltranslation/archive/master.zip
https://bitbucket.org/izi/django-admin-tools/downloads/django-admin-tools-0.5.1.tar.gz
easy-thumbnails==1.2
psycopg2==2.5
django-tinymce==1.5.1
ipython==0.13.2
django-autoslug==1.6.1
django-embed-video==0.2.0
requests==1.2.3
